pyproject.toml
[project]
name = "tg-talking-agent"
version = "0.1.0"
description = "Telegram AI Agent that mimics your communication style"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.34.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "InquirerPy>=0.3.4",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
tg-agent = "src.main:main"
